Further clean up and prepare for UI editor
Removed more stuff I don't use and created a deprecated directory for
things I might want to restore in the future (somehow). Automations have
been moved to automations.yaml to support the config UI, but it's quite
useless until templates (e.g. service_template) are supported. Will move
scripts and customize as well.
