+++ homeassistant/configuration.yaml
@@ -10,6 +10,8 @@ homeassistant:
 
   packages: !include_dir_named packages
 
+automation: !include automations.yaml
+
 logbook:
   exclude:
     entities:
