+++ homeassistant/custom_components/binary_sensor/dlink.py
@@ -7,13 +7,16 @@ import urllib
 import logging
 import asyncio
 import functools
+import aiohttp
+import xml.etree.ElementTree as ET
+
+from io import BytesIO
 from datetime import datetime
 
-from pysimplesoap.client import SoapClient
+import xmltodict
 
 _LOGGER = logging.getLogger(__name__)
 
-NAMESPACE = 'http://purenetworks.com/HNAP1/'
 ACTION_BASE_URL = 'http://purenetworks.com/HNAP1/'
 
 
@@ -31,14 +34,14 @@ class AuthenticationError(Exception):
 class HNAPClient:
     """Client for the HNAP protocol."""
 
-    def __init__(self, address, username, password, loop=None):
+    def __init__(self, soap, username, password, loop=None):
         """Initialize a new HNAPClient instance."""
-        self.address = address
         self.username = username
         self.password = password
         self.logged_in = False
         self.loop = loop or asyncio.get_event_loop()
         self.actions = None
+        self._client = soap
         self._private_key = None
         self._cookie = None
         self._auth_token = None
@@ -47,14 +50,15 @@ class HNAPClient:
     @asyncio.coroutine
     def login(self):
         """Authenticate with device and obtain cookie."""
+        _LOGGER.info('Logging into device')
         self.logged_in = False
         resp = yield from self.call(
             'Login', Action='request', Username=self.username,
             LoginPassword='', Captcha='')
 
-        challenge = str(resp.LoginResponse.Challenge)
-        public_key = str(resp.LoginResponse.PublicKey)
-        self._cookie = str(resp.LoginResponse.Cookie)
+        challenge = resp['Challenge']
+        public_key = resp['PublicKey']
+        self._cookie = resp['Cookie']
         _LOGGER.debug('Challenge: %s, Public key: %s, Cookie: %s',
                       challenge, public_key, self._cookie)
 
@@ -67,7 +71,7 @@ class HNAPClient:
                 'Login', Action='login', Username=self.username,
                 LoginPassword=password, Captcha='')
 
-            if str(resp.LoginResult).lower() != 'success':
+            if resp['LoginResult'].lower() != 'success':
                 raise AuthenticationError('Incorrect username or password')
 
             if not self.actions:
@@ -80,11 +84,9 @@ class HNAPClient:
 
     @asyncio.coroutine
     def device_actions(self):
-        def _extract(action):
-            url = str(action)
-            return url[url.rfind('/')+1:]
         actions = yield from self.call('GetDeviceSettings')
-        return list(map(_extract, actions.SOAPActions.children()))
+        return list(map(lambda x: x[x.rfind('/')+1:],
+                        actions['SOAPActions']['string']))
 
     @asyncio.coroutine
     def soap_actions(self, module_id):
@@ -94,33 +96,23 @@ class HNAPClient:
     @asyncio.coroutine
     def call(self, method, *args, **kwargs):
         """Call an NHAP method (async)."""
-        def _call_method():
-            self._update_nauth_token(method)
-            method_to_call = getattr(self._client(), method)
-            return self.loop.run_in_executor(
-                None, functools.partial(method_to_call, *args, **kwargs))
-
         # Do login if no login has been done before
         if not self._private_key and method != 'Login':
             yield from self.login()
 
+        self._update_nauth_token(method)
         try:
-            res = yield from _call_method()
+            result = yield from self.soap().call(method, **kwargs)
+            if 'ERROR' in result:
+                self._bad_response()
+        except:
+            self._bad_response()
+        return result
 
-            # Force login if if we a HTTP page (it's likely Unauthorized)
-            if 'body' not in res:
-                return res
-
-        except AttributeError:
-            _LOGGER.debug('Got logged out, logging in again')
-        except xml.parsers.expat.ExpatError:
-            _LOGGER.debug('Got logged out, logging in again')
-        except urllib.error.HTTPError:
-            _LOGGER.debug('Got logged out, logging in again')
-        except urllib.error.URLError:
-            _LOGGER.debug('Timed out')
-        yield from self.login()
-        return (yield from _call_method())
+    def _bad_response(self):
+        _LOGGER.error('Got an error, resetting private key')
+        self._private_key = None
+        raise Exception('got error response from device')
 
     def _update_nauth_token(self, action):
         """Update NHAP auth token for an action."""
@@ -134,30 +126,15 @@ class HNAPClient:
         _LOGGER.debug('Generated new token for %s: %s (time: %d)',
                       action, self._auth_token, self._timestamp)
 
-    def _client(self):
-        headers = {}
-
+    def soap(self):
+        """Get SOAP client with updated headers."""
         if self._cookie:
-            headers['Cookie'] = 'uid={0}'.format(self._cookie)
+            self._client.headers['Cookie'] = 'uid={0}'.format(self._cookie)
         if self._auth_token:
-            headers['HNAP_AUTH'] = '{0} {1}'.format(
+            self._client.headers['HNAP_AUTH'] = '{0} {1}'.format(
                 self._auth_token, self._timestamp)
 
-        client = SoapClient(
-            'http://{0}/HNAP1'.format(self.address), trace=False,
-            namespace=NAMESPACE,
-            action=ACTION_BASE_URL,
-            http_headers=headers)
-
-        # The device is very picky about the request and will fail if an empty
-        # Header-tag is included. Unfortunately, it's not possible to remove
-        # that tag from pysimplesoap without modifying the library. This is a
-        # hack that modifies the internal request and removes the header. Not
-        # pretty, but will have to do for now. Over and out.
-        # pylint: disable=protected-access
-        client._SoapClient__xml = client._SoapClient__xml.replace(
-            '<%(soap_ns)s:Header/>\n', '')
-        return client
+        return self._client
 
 
 class MotionSensor:
@@ -179,79 +156,83 @@ class MotionSensor:
         if 'GetLatestDetection' in self._soap_actions:
             resp = yield from self.client.call(
                 'GetLatestDetection', ModuleID=self.module_id)
-            detect_time = float(resp.LatestDetectTime)
+            detect_time = resp['LatestDetectTime']
         else:
             resp = yield from self.client.call(
                 'GetMotionDetectorLogs', ModuleID=self.module_id, MaxCount=1,
                 PageOffset=1, StartTime=0, EndTime='All')
-            detect_time = float(resp.MotionDetectorLogList[0].TimeStamp)
+            if 'MotionDetectorLogList' not in resp:
+                _LOGGER.error('log list: ' + str(resp))
+            log_list = resp['MotionDetectorLogList']
+            detect_time = log_list['MotionDetectorLog']['TimeStamp']
 
-        return datetime.fromtimestamp(detect_time)
-
-    @asyncio.coroutine
-    def loop(self):
-        latest = None
-        while True:
-            try:
-                trigger = yield from self.latest_trigger()
-                if latest != trigger:
-                    print(trigger)
-                    latest = trigger
-            except Exception as ex:
-                print("EXCEPTION: " + str(ex))
-            yield from asyncio.sleep(5)
-
-    @asyncio.coroutine
-    def module_actions(self):
-        resp = yield from self.client.call(
-            'GetModuleSOAPActions', ModuleID=self.module_id)
-        print(resp)
-
-    @asyncio.coroutine
-    def profile(self):
-        resp = yield from self.client.call(
-            'GetModuleProfile', ModuleID=self.module_id)
-        print(resp)
+        return datetime.fromtimestamp(float(detect_time))
 
     @asyncio.coroutine
     def system_log(self):
         resp = yield from self.client.call(
-            'GetSystemLogs', MaxCount=100, Tag='All',
+            'GetSystemLogs', MaxCount=100,
             PageOffset=1, StartTime=0, EndTime='All')
-        print(resp)
-
-    @asyncio.coroutine
-    def firmware_status(self):
-        resp = yield from self.client.call('GetFirmwareStatus')
-        print(resp)
-
-    @asyncio.coroutine
-    def internet_status(self):
-        resp = yield from self.client.call('GetCurrentInternetStatus')
-        print(resp)
-
-    @asyncio.coroutine
-    def internet_settings(self):
-        resp = yield from self.client.call('GetInternetSettings')
         print(resp)
 
     @asyncio.coroutine
     def _cache_soap_actions(self):
         resp = yield from self.client.soap_actions(self.module_id)
-        actions = filter(lambda x: x.get_name() == 'Action',
-                         resp.SOAPActions.children())
-        self._soap_actions = list(map(lambda x: str(x), actions))
+        self._soap_actions = resp['ModuleSOAPList']['SOAPActions']['Action']
 
-    # This is for siren
+
+class NanoSOAPClient:
+
+    BASE_NS = {'xmlns:soap': 'http://schemas.xmlsoap.org/soap/envelope/',
+               'xmlns:xsd': 'http://www.w3.org/2001/XMLSchema',
+               'xmlns:xsi': 'http://www.w3.org/2001/XMLSchema-instance'}
+    ACTION_NS = {'xmlns': 'http://purenetworks.com/HNAP1/'}
+
+    def __init__(self, address, action, loop=None, session=None):
+        self.address = 'http://{0}/HNAP1'.format(address)
+        self.action = action
+        self.loop = loop or asyncio.get_event_loop()
+        self.session = session or aiohttp.ClientSession(loop=loop)
+        self.headers = {}
+
+    def _generate_request_xml(self, method, **kwargs):
+        body = ET.Element('soap:Body')
+        action = ET.Element(method, self.ACTION_NS)
+        body.append(action)
+
+        for param, value in kwargs.items():
+            element = ET.Element(param)
+            element.text = str(value)
+            action.append(element)
+
+        envelope = ET.Element('soap:Envelope', self.BASE_NS)
+        envelope.append(body)
+
+        f = BytesIO()
+        tree = ET.ElementTree(envelope)
+        tree.write(f, encoding='utf-8', xml_declaration=True)
+
+        return f.getvalue().decode('utf-8')
+
     @asyncio.coroutine
-    def sound_play(self, sound_type, volume, duration, controller):
-        resp = yield from self.client.call(
-            'SetSoundPlay', ModuleID=self.module_id,
-            SoundType=sound_type, Volume=volume,
-            Duration=duration, Controller=controller)
-        print(resp)
+    def call(self, method, **kwargs):
+        xml = self._generate_request_xml(method, **kwargs)
+
+        headers = self.headers.copy()
+        headers['SOAPAction'] = '"{0}{1}"'.format(self.action, method)
+
+        resp = yield from self.session.post(
+            self.address, data=xml, headers=headers, timeout=10)
+        text = yield from resp.text()
+        parsed = xmltodict.parse(text)
+        if 'soap:Envelope' not in parsed:
+            _LOGGER.error("parsed: " + str(parsed))
+            raise Exception('probably a bad response')
+
+        return parsed['soap:Envelope']['soap:Body'][method + 'Response']
 
 if __name__ == '__main__':
+    logging.basicConfig(level=logging.DEBUG)
     loop = asyncio.get_event_loop()
 
     import sys
@@ -261,7 +242,10 @@ if __name__ == '__main__':
 
     @asyncio.coroutine
     def _print_latest_motion():
-        client = HNAPClient(address, 'Admin', pin, loop=loop)
+        session = aiohttp.ClientSession()
+        soap = NanoSOAPClient(
+            address, ACTION_BASE_URL, loop=loop, session=session)
+        client = HNAPClient(soap, 'Admin', pin, loop=loop)
         motion = MotionSensor(client)
         yield from client.login()
 
@@ -271,22 +255,9 @@ if __name__ == '__main__':
         elif cmd == 'actions':
             print('Supported actions:')
             print('\n'.join(client.actions))
-        elif cmd == 'system_log':
-            yield from motion.system_log()
-        elif cmd == 'module_actions':
-            yield from motion.module_actions()
-        elif cmd == 'profile':
-            yield from motion.profile()
-        elif cmd == 'sound_play':
-            yield from motion.sound_play(
-                sys.argv[4], sys.argv[5], sys.argv[6], sys.argv[7])
-        elif cmd == 'firmware_status':
-            yield from motion.firmware_status()
-        elif cmd == 'internet_status':
-            yield from motion.internet_status()
-        elif cmd == 'internet_settings':
-            yield from motion.internet_settings()
-        elif cmd == 'loop':
-            yield from motion.loop()
+        elif cmd == 'log':
+            log = yield from motion.system_log()
+
+        session.close()
 
     loop.run_until_complete(_print_latest_motion())
