+++ homeassistant/deprecated/camera.yaml
