--- homeassistant/packages/automations/auto_sonos.yaml
@@ -1,56 +1,0 @@
-input_boolean:
-  sonos_run_today:
-    name: Sonos automation has run today
-
-automation:
-  - alias: Reset sonos_run_today
-    trigger:
-      platform: time
-      at: '02:00:00'
-    action:
-      service: input_boolean.turn_off
-      entity_id: input_boolean.sonos_run_today
-
-  - alias: Turn on Sonos when someone comes home from work
-    trigger:
-      platform: state
-      entity_id: binary_sensor.kitchen_sensor
-      from: 'off'
-    condition:
-      - condition: state
-        entity_id: input_boolean.sonos_run_today
-        state: 'off'
-      - condition: state
-        entity_id: group.family
-        state: home
-      - condition: time
-        after: '15:00:00'
-        weekday:
-        - fri
-    action:
-      - service: input_boolean.turn_on
-        entity_id: input_boolean.sonos_run_today
-      - service: script.sonos_say
-        data:
-          sonos_entity: media_player.kok
-          message: Det är fredag, dags för lite musik!
-          delay: '00:00:05'
-          volume: 0.5
-      - delay: '00:00:07'
-      - service: media_player.media_play
-        entity_id: media_player.kok
-
-  - alias: Turn off bedroom light when emby plays
-    trigger:
-      platform: state
-      entity_id: media_player.emby__emby_mobile__chrome
-      to: 'playing'
-    condition:
-      - condition: state
-        entity_id: group.family
-        state: home
-      - condition: time
-        after: '21:00:00'
-    action:
-      - service: switch.turn_off
-        entity_id: switch.pierre_bedroom
