--- homeassistant/packages/automations/lights.yaml
@@ -1,55 +1,0 @@
-# Turns on bedroom light if:
-# - Someone is home
-# - Motion is detected in bedroom
-# - Time is after 21:30
-# - Only once per day
-input_boolean:
-  pierre_bedroom_run_today:
-    name: Turn on Pierres bedroom lamp automation has run today
-
-automation:
-  - alias: Reset pierre_bedroom_run_today
-    trigger:
-      platform: time
-      at: '08:00:00'
-    action:
-      service: input_boolean.turn_off
-      entity_id: input_boolean.pierre_bedroom_run_today
-
-  - alias: Turn on bedroom lamp when someone enters the room
-    trigger:
-      platform: state
-      entity_id: binary_sensor.bedroom_sensor
-      from: 'off'
-    condition:
-      - condition: state
-        entity_id: input_boolean.pierre_bedroom_run_today
-        state: 'off'
-      - condition: state
-        entity_id: group.family
-        state: home
-      - condition: time
-        after: '21:30:00'
-    action:
-      - service: input_boolean.turn_on
-        entity_id: input_boolean.pierre_bedroom_run_today
-      - service: switch.turn_on
-        entity_id: switch.pierre_bedroom
-
-  # TODO: create a scene for "evning lights" and use that instead
-  - alias: Turn on evening lights
-    trigger:
-      - platform: sun
-        event: sunset
-        offset: '00:15:00'
-      - platform: sun
-        event: sunrise
-        offset: '01:00:00'
-    action:
-      service_template: >
-        {% if trigger.event == 'sunset' %}
-          switch.turn_on
-        {% else %}
-          switch.turn_off
-        {% endif %}
-      entity_id: ['switch.downstair_lights', 'switch.upstair_lights']
