--- homeassistant/packages/receiver.yaml
@@ -1,5 +1,0 @@
-media_player:
-  - platform: yamaha
-    name: Receiver
-    zone_ignore:
-      - Zone_2
