+++ homeassistant/packages/sonos.yaml
@@ -12,6 +12,7 @@ switch:
     switches:
       sonos_player:
         friendly_name: Sonos
+        entity_id: media_player.kok
         value_template: "{{ is_state('media_player.kok', 'playing') }}"
         turn_on:
           service: media_player.media_play
