--- homeassistant/packages/switches/mqtt.yaml
@@ -1,10 +1,0 @@
-# Dummy switches for now (not used)
-switch:
-  - platform: mqtt
-    name: Pierre Bedroom
-    command_topic: cmnd/pierre_bedroom/power
-    state_topic: stat/pierre_bedroom/POWER
-  - platform: mqtt
-    name: mailbox
-    command_topic: sensor/mailBox
-    state_topic: sensor/mailBox
